Add provider service and deprecate legacy runtime interface (#19)
* Add new provider service and driver

* Cleanup unused code from legacy runtime interface
