+++ initializer.go
@@ -7,10 +7,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/unweave/unweave/api/types"
 	"github.com/unweave/unweave/builder"
-	"github.com/unweave/unweave/providers/lambdalabs"
-	"github.com/unweave/unweave/runtime"
 	"github.com/unweave/unweave/tools/gonfig"
 	"github.com/unweave/unweave/vault"
 )
@@ -24,31 +21,6 @@ type providerConfig struct {
 
 type builderConfig struct {
 	RegistryURI string `env:"UNWEAVE_CONTAINER_REGISTRY_URI"`
-}
-
-func (i *EnvInitializer) InitializeRuntime(ctx context.Context, userID string, provider types.Provider) (*runtime.Runtime, error) {
-	var cfg providerConfig
-	gonfig.GetFromEnvVariables(&cfg)
-
-	switch provider {
-	case types.LambdaLabsProvider:
-		if cfg.LambdaLabsAPIKey == "" {
-			return nil, fmt.Errorf("missing LambdaLabs API key in runtime config file")
-		}
-		node, err := lambdalabs.NewNodeRuntime(cfg.LambdaLabsAPIKey)
-		if err != nil {
-			return nil, err
-		}
-		sess, err := lambdalabs.NewSessionRuntime(cfg.LambdaLabsAPIKey)
-		if err != nil {
-			return nil, err
-		}
-
-		return &runtime.Runtime{Node: node, Exec: sess}, nil
-
-	default:
-		return nil, fmt.Errorf("%q provider not supported in the env initializer", provider)
-	}
 }
 
 func (i *EnvInitializer) InitializeBuilder(ctx context.Context, userID string, builderType string) (builder.Builder, error) {
